--- backend/src/main/java/com/bungeobbang/backend/login/dto/request/LoginRequest.java
@@ -1,4 +1,0 @@
-package com.bungeobbang.backend.login.dto.request;
-
-public record LoginRequest(String code) {
-}
