+++ backend/src/main/java/com/bungeobbang/backend/member/domain/MemberTokens.java
@@ -1,4 +1,4 @@
-package com.bungeobbang.backend.login.domain;
+package com.bungeobbang.backend.member.domain;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
