+++ backend/src/main/java/com/bungeobbang/backend/member/domain/OauthAccessToken.java
@@ -1,4 +1,4 @@
-package com.bungeobbang.backend.login.domain;
+package com.bungeobbang.backend.member.domain;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Getter;
