+++ backend/src/main/java/com/bungeobbang/backend/member/domain/OauthUserInfo.java
@@ -1,4 +1,4 @@
-package com.bungeobbang.backend.login.domain;
+package com.bungeobbang.backend.member.domain;
 
 public interface OauthUserInfo {
     String getSocialLoginId();
