+++ backend/src/main/java/com/bungeobbang/backend/member/dto/request/LoginRequest.java
@@ -1,0 +1,4 @@
+package com.bungeobbang.backend.member.dto.request;
+
+public record LoginRequest(String code) {
+}
