+++ backend/src/main/java/com/bungeobbang/backend/member/dto/response/LoginResponse.java
@@ -1,4 +1,4 @@
-package com.bungeobbang.backend.login.dto.response;
+package com.bungeobbang.backend.member.dto.response;
 
 public record LoginResponse(Long memberId, Boolean isEmailVerified, String accessToken, String refreshToken) {
 }
