+++ backend/src/main/java/com/bungeobbang/backend/member/presentation/MemberController.java
@@ -1,7 +1,7 @@
-package com.bungeobbang.backend.login.presentation;
+package com.bungeobbang.backend.member.presentation;
 
-import com.bungeobbang.backend.login.dto.response.LoginResponse;
-import com.bungeobbang.backend.login.service.LoginService;
+import com.bungeobbang.backend.member.dto.response.LoginResponse;
+import com.bungeobbang.backend.member.service.MemberService;
 import com.bungeobbang.backend.member.domain.ProviderType;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
@@ -9,9 +9,10 @@ import org.springframework.web.bind.annotation.*;
 
 @RestController
 @RequiredArgsConstructor
-public class LoginController {
+public class MemberController {
 
     private final LoginService loginService;
+    private final MemberService memberService;
 
     @GetMapping("/student/auth/{provider}/login")
     public ResponseEntity<LoginResponse> login(
@@ -19,7 +20,7 @@ public class LoginController {
             @RequestParam final String code
             ) {
         final ProviderType providerType = ProviderType.fromString(provider);
-        final LoginResponse loginResponse = loginService.login(providerType, code);
+        final LoginResponse loginResponse = memberService.login(providerType, code);
         return ResponseEntity.ok(loginResponse);
     }
 }
